register: Use autofocus attribute to focus the first input.
There are two reason to do this approach: it's cleaner and it fixes a
puppeteer flake. With the autofocus attribute the browser will focus
the input as soon as it can. The common.autofocus code we used previously
wraps the focus code in $() which occurs after DOMContentLoaded or load
event at which point the input would have been already useable thus focused.
This change probably predates the `defer` attribute we add to script tags so
the script runs after the dom is ready which would have made this the use
of $() nessacary because we load js in head before the input.

The puppeteer flake will occur with previous code is that it will be typing
something in a field and in the middle of it the previous code will focus
the first input causing most of the input to be typed into the first input.
That leads to validation errors and failures that occur nondeterministically.

Also, we remove the use of common.autofocus here because the whole file
is already contained in a $() call which the common.autofocus wraps the focus
code in does again. Instead we just call the focus method to clean it up.
The comments are also tweaked a bit too.