puppetter: Fix test flake in 01-login.js.
There were two reason there were flakes in login tests. One problem was
that the assert call for checking that the login page url was happening
too early in rare cases. To fix this issue we wait for email input field
element in the login page. The second issue was the sometimes the url was
/login/ instead of /accounts/login/. The time between the redirect were
super close here too (around ~10ms) so I am suprised this didn't occur flake
locally more often.

Server logs that show the redirect:
2020-05-26 18:55:27.065 INFO [zr] 127.0.0.1       GET     200  36ms (db: 5ms/2q) (+start: 14ms) /accounts/login/ (unauth@zulip via ?)
2020-05-26 18:55:27.074 INFO [zr] 127.0.0.1       GET     200  28ms (db: 3ms/2q) (+start: 23ms) /login/ (unauth@zulip via ?)