ts: Convert util to typescript.
The only code changes in this commits is to call .toString in the
clean_user_content_links function because elt.setAttribute expects a
string. The browser does this conversation automatically. And adding
a number check in default less function which is reduadent because
ensure it is a number before but needed and harmless. Finally, we
remove the unassigned_value_sentinal in favor of an enum type.

We also update exclude rule in custom_rule.py to use util.ts and
update the langs to add js. This is no other place that needs ts to
be added in lang along with js.