github-actions: Rename tools/circleci -> tools/github_actions.
Along, with the directory rename and references updates. We also
update the Dockerfile.template to create a github user instead of
circleci.

Note: references in .circleci/config.yml and docs are not updated,
since nothing everything is fully migrated yet. I expect remove the
circleci config, and update docs when we are done with the
migration.