github-actions: Add bionic production build job.
All the steps are same from circleci expect the Add required
permissions and presist_to_workspace. Permissions step is github
actions specific as explained in comments. presist_to_workspace is
equivlent to upload-artifact in GitHub Actions. Since the upload
artificat doesn't allow us to upload select files under single
artificat we prepare a directory with files we want to upload
using the Prepare production tarball directory step.

Finally, I should note the duplication in this and zulip-ci
worklflow. I don't think it would be cleaner to try to mush this
workflow there and we plan to only run this when production related
directorires are updated so we need to split it to its own
workspace. We didn't update any step since they were added so at
least that's good.