github-actions: Add puppeteer stress test workflows.
This run 4 jobs that run 100 tests each whenever puppeteer tests
or this workflow is updated. It does not run on zulip/zulip pushes
to avoid backlog caused by the long time this will take to finish;
it runs on pull requests and pushes to fork. This just make looking
triggering these tests easier since previously we have to push a new
branch and update the CI config to accomplish this.

All the steps are from zulip-ci workflow with expection to last two
new steps and the cache key uses github.job (which is stress_tests)
in cache key rather than matrix.os.