puppeteer: Fix flake with compose box input.
There were a lots of flakes in CI recently because typhead didn't
appear when Enter was pressed and real emails are not accepted as
valid inputs. To fix this we wait for typhead to appear and then
click that instead of Enter.

Re-enable puppeteer test in CI after this fix too.