puppeteer: Add code to record tests using chrome extensions.
This code is borrowed from https://github.com/muralikg/puppetcam and
has been heavily modified to fit our needs. Most of that code was
outdated and has no API structure to it; it was meant to used a CLI
and on top of that it offered no control to when to stop recording.
It waited for couple of seconds and then quit :(

We update almost every single aspect of the code to make it work.
At some point, I plan to move this out to a npm package.