dependencies: Upgrade babel-plugin-rewire-ts to v1.4.0.
This version keeps both the internal and external state of the ES6
module in sync so it's state is  similar to a commonjs module.
Since, we don't need to ask rewire for the internal state of the
module the tests will be same as they were before the TS/ES6/Remove
from window migrations.