node fixtures: Clean up realm_emoji data.
We extract test_realm_emojis, and we make
the name of the event more explicit (adding
the __update suffix).

We also add the "op" of "update" here, which
is sort of a quirk of the api, since we don't
actually have alternatives like add/remove,
and therefore the current frontend code doesn't
look at the "op", and thus the original tests
never had to provide a correct value for it.