node_tests: Don't remove require cache of module in zrequire.
There is good reason to do this (explanation is bit long!). With the
TypeScript migration, and the require and ES6 migrations that come
with it, we use require instead of set_global which loads the entire
module. Suppose we have util module, which is used by some other
module, say message_store and util is being required in
message_store since it removed from window. Then, if an test
zrequire's message_store first and then zrequire's the util module
and mocks one of it methods it will not be mocked for the
message_store module. The reason is:
 1. zrequire('message_store') leads to require('util').
 2. zrequire('util') removes the util module from cache and it is
    reloaded. Now the util module in message_store and the one in
    test will be diffrent and any updates to it in tests won't
    reflect in the actual code.

Which can leads to confusion for folks writing tests. I'll mention
this can be avoided doing zrequire('util') first but...that is not
ideal.

And, since there was one outlier test that relied on this behavoir
we add namespace.reset_module function.